Move the `delay` utility function to the async module
