+++ src/async/validation.test.ts
@@ -15,8 +15,9 @@
  */
 
 import { describe, expect, it } from "vitest";
-import { Str, TestBuilder, delay } from "../_test/utils.js";
+import { Str, TestBuilder } from "../_test/utils.js";
 import { Validation } from "../validation.js";
+import { delay } from "./_test/utils.js";
 import { AsyncValidation } from "./validation.js";
 
 describe("AsyncValidation", () => {
